Added setup.py installation functionality
